@@ -25,8 +25,8 @@
   "devDependencies": {
     "@fluentui/react-northstar": "0.61.0",
     "@microsoft/api-extractor": "~7.18.0",
-    "@types/react": "^16.9.49",
-    "@types/react-dom": "^16.9.8",
+    "@types/react": "18.2.0",
+    "@types/react-dom": "18.2.1",
     "@typescript-eslint/parser": "^4.12.0",
     "eslint": "^7.7.0",
     "eslint-config-prettier": "^6.12.0",
@@ -34,8 +34,8 @@
     "eslint-plugin-header": "^3.1.0",
     "eslint-plugin-import": "~2.22.1",
     "prettier": "2.3.1",
-    "react": "~16.14.0",
-    "react-dom": "16.13.1",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "scheduler": "^0.20.0",
     "rimraf": "^2.6.2",
     "typescript": "3.8.3",
