@@ -44,8 +44,8 @@
     "@internal/react-composites": "1.5.1-beta.4",
     "merge": "^2.1.1",
     "mobile-detect": "~1.4.5",
-    "react": "~16.14.0",
-    "react-dom": "16.13.1",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-router-dom": "^5.1.2",
     "reselect": "~4.0.0",
     "uuid": "^8.1.0"
@@ -53,11 +53,10 @@
   "devDependencies": {
     "@babel/core": "~7.16.0",
     "@babel/preset-env": "7.13.10",
-    "@testing-library/react-hooks": "^3.4.2",
     "@types/jest": "~26.0.22",
     "@types/node": "^14.14.10",
-    "@types/react": "^16.9.49",
-    "@types/react-dom": "^16.9.8",
+    "@types/react": "18.2.0",
+    "@types/react-dom": "18.2.1",
     "@types/uuid": "^8.3.0",
     "@typescript-eslint/eslint-plugin": "^4.12.0",
     "@typescript-eslint/parser": "^4.12.0",
@@ -87,7 +86,6 @@
     "jest-junit": "~13.0.0",
     "prettier": "2.3.1",
     "pretty-quick": "^3.1.0",
-    "react-test-renderer": "^16.14.0",
     "rimraf": "^2.6.2",
     "source-map-explorer": "^2.5.0",
     "style-loader": "~2.0.0",
