@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 import React from 'react';
-import { renderHook } from '@testing-library/react-hooks';
+import { renderHook } from '@testing-library/react';
 import { _PermissionsProvider, _getPermissions, _Permissions, _usePermissions } from './PermissionsProvider';
 
 const mockPresenterPermissions: _Permissions = {
