@@ -3,8 +3,8 @@
 
 import {
   sendMessageFromHiddenChatComposite,
-  startTypeMessageFromHiddenChatComposite,
-  stopTypingAndSendMessageFromHiddenChatComposite,
+  typeMessageFromHiddenChatComposite,
+  typeAndSendMessageFromHiddenChatComposite,
   temporarilyShowHiddenChatComposite
 } from '../../common/hermeticChatTestHelpers';
 import { defaultMockCallAdapterState, defaultMockRemoteParticipant } from '../../call/hermetic/fixture';
@@ -42,11 +42,11 @@ test.describe('CallWithChat Composite CallWithChat Page Tests', () => {
     await waitForMessageSeen(page);
 
     // Start typing a message
-    await startTypeMessageFromHiddenChatComposite(page, chatRemoteParticipant, 'I agree!');
+    await typeMessageFromHiddenChatComposite(page, chatRemoteParticipant, 'I agree!');
     // Wait for indicator to show up
     await waitForAndHideTypingIndicator(page, APP_UNDER_TEST_ROOT_SELECTOR);
     // Stop typing and send a message
-    await stopTypingAndSendMessageFromHiddenChatComposite(page, chatRemoteParticipant);
+    await typeAndSendMessageFromHiddenChatComposite(page, chatRemoteParticipant, 'I agree!');
 
     // Local participant has both a sent message and a received message.
     await waitForNMessages(page, 2, APP_UNDER_TEST_ROOT_SELECTOR);
