@@ -5,7 +5,7 @@ import type { ChatParticipant } from '@azure/communication-chat';
 import { toFlatCommunicationIdentifier } from '@internal/acs-ui-common';
 import { Page } from '@playwright/test';
 import { IDS } from './constants';
-import { dataUiId, perStepLocalTimeout, screenshotOnFailure } from './utils';
+import { dataUiId, perStepLocalTimeout, screenshotOnFailure, waitForSelector } from './utils';
 
 /**
  * <HiddenChatComposites /> are, well, hidden.
@@ -15,7 +15,10 @@ import { dataUiId, perStepLocalTimeout, screenshotOnFailure } from './utils';
  */
 export async function temporarilyShowHiddenChatComposite(page: Page, participant: ChatParticipant): Promise<void> {
   await withHiddenChatCompositeInForeground(page, participant, async () => {
-    return;
+    // wait for messages to have loaded
+    await waitForSelector(page, dataUiId('chat-composite-message'));
+    // sleep for 100ms to give time for the messagethread hook to send the read receipt
+    await page.waitForTimeout(100);
   });
 }
 
@@ -35,6 +38,9 @@ export async function withHiddenChatCompositeInForeground(
     // Temporarily bring chat composite to foreground.
     await handle.evaluate((node) => (node.style.display = 'block'));
     await handle.waitFor({ state: 'visible', timeout: perStepLocalTimeout() });
+
+    // wait for composite to have loaded
+    await waitForSelector(page, dataUiId(IDS.sendboxTextField));
   });
   await action();
   await screenshotOnFailure(page, async () => {
@@ -63,11 +69,14 @@ export async function sendMessageFromHiddenChatComposite(
 }
 
 /**
- * Start Type a message from one of the hidden chat composites
+ * Type a message from one of the hidden chat composites. This is useful for testing the typing indicator.
+ *
+ * @remarks
+ * Becuase the hidden chat composite is unmounted when hidden, anything typed here is lost when the composite is hidden.
  *
  * @private
  */
-export async function startTypeMessageFromHiddenChatComposite(
+export async function typeMessageFromHiddenChatComposite(
   page: Page,
   participant: ChatParticipant,
   message: string
@@ -80,15 +89,19 @@ export async function startTypeMessageFromHiddenChatComposite(
 }
 
 /**
- * Stop typing and send the message from one of the hidden chat composites.
+ * Type and send the message from one of the hidden chat composites.
  *
  * @private
  */
-export async function stopTypingAndSendMessageFromHiddenChatComposite(
+export async function typeAndSendMessageFromHiddenChatComposite(
   page: Page,
-  participant: ChatParticipant
+  participant: ChatParticipant,
+  message: string
 ): Promise<void> {
   await withHiddenChatCompositeInForeground(page, participant, async () => {
+    await page.type(`${hiddenCompositeSelector(participant)} ${dataUiId(IDS.sendboxTextField)}`, message, {
+      timeout: perStepLocalTimeout()
+    });
     await page.focus(`${hiddenCompositeSelector(participant)} ${dataUiId(IDS.sendboxTextField)}`);
     await page.keyboard.press('Enter');
   });
